+++ server/src/test/java/org/realityforge/gwt/appcache/server/PermutationTest.java
@@ -1,6 +1,6 @@
 package org.realityforge.gwt.appcache.server;
 
-import java.util.HashSet;
+import java.util.ArrayList;
 import org.testng.annotations.Test;
 import static org.testng.Assert.*;
 
@@ -11,15 +11,14 @@ public class PermutationTest
     throws Exception
   {
     final Permutation permutation1 = new Permutation( "Foo" );
-    final Permutation permutation2 = new Permutation( "Bar" );
     assertEquals( permutation1.getPermutationName(), "Foo" );
-    assertEquals( permutation1.getBindingProperties().size(), 0 );
-    permutation1.getBindingProperties().put( 0, new HashSet<BindingProperty>() );
-    assertEquals( permutation1.getBindingProperties().size(), 1 );
+    assertEquals( permutation1.getSelectors().size(), 0 );
+    permutation1.getSelectors().add( new SelectionDescriptor( "Foo", new ArrayList<BindingProperty>() ) );
+    assertEquals( permutation1.getSelectors().size(), 1 );
     assertEquals( permutation1.getPermutationFiles().size(), 0 );
     permutation1.getPermutationFiles().add( "foo.txt" );
     assertEquals( permutation1.getPermutationFiles().size(), 1 );
-    assertFalse( permutation1.equals( permutation2 ) );
+    assertFalse( permutation1.equals( new Permutation( "Bar" ) ) );
     assertTrue( permutation1.equals( permutation1 ) );
   }
 }